+++ conjure-java-core/src/main/java/com/palantir/conjure/java/types/BeanGenerator.java
@@ -24,8 +24,6 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
-import com.google.common.collect.Iterators;
-import com.google.common.collect.PeekingIterator;
 import com.palantir.conjure.CaseConverter;
 import com.palantir.conjure.java.ConjureAnnotations;
 import com.palantir.conjure.java.Options;
@@ -39,7 +37,6 @@ import com.palantir.conjure.java.visitor.MoreVisitors;
 import com.palantir.conjure.spec.FieldDefinition;
 import com.palantir.conjure.spec.FieldName;
 import com.palantir.conjure.spec.ListType;
-import com.palantir.conjure.spec.LogSafety;
 import com.palantir.conjure.spec.MapType;
 import com.palantir.conjure.spec.ObjectDefinition;
 import com.palantir.conjure.spec.OptionalType;
@@ -63,13 +60,10 @@ import com.squareup.javapoet.TypeSpec;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import javax.annotation.Nonnull;
 import javax.lang.model.element.Modifier;
 import org.apache.commons.lang3.StringUtils;
 import org.immutables.value.Value;
@@ -100,10 +94,6 @@ public final class BeanGenerator {
                 Packages.getPrefixedName(typeDef.getTypeName(), options.packagePrefix());
         ClassName objectClass = ClassName.get(prefixedName.getPackage(), prefixedName.getName());
         ClassName builderClass = ClassName.get(objectClass.packageName(), objectClass.simpleName(), "Builder");
-        ClassName builderImplementation =
-                options.useStagedBuilders() && fields.stream().anyMatch(field -> !fieldShouldBeInFinalStage(field))
-                        ? ClassName.get(objectClass.packageName(), objectClass.simpleName(), "DefaultBuilder")
-                        : builderClass;
 
         // Use the fully-resolved/computed safety value for the top-level class, however
         // fields (setters/getters) are annotated with declared values because complex objects
@@ -158,13 +148,17 @@ public final class BeanGenerator {
                         .build());
             }
         } else {
+            ClassName builderImplementation = options.useStagedBuilders()
+                            && fields.stream().anyMatch(field -> !BeanBuilderGenerator.fieldShouldBeInFinalStage(field))
+                    ? ClassName.get(objectClass.packageName(), objectClass.simpleName(), "DefaultBuilder")
+                    : builderClass;
             ImmutableList<EnrichedField> fieldsNeedingBuilderStage = fields.stream()
-                    .filter(field -> !fieldShouldBeInFinalStage(field))
+                    .filter(field -> !BeanBuilderGenerator.fieldShouldBeInFinalStage(field))
                     .collect(ImmutableList.toImmutableList());
-            typeBuilder.addAnnotation(AnnotationSpec.builder(JsonDeserialize.class)
-                    .addMember("builder", "$T.class", builderImplementation)
-                    .build());
             if (!options.useStagedBuilders() || fieldsNeedingBuilderStage.isEmpty()) {
+                typeBuilder.addAnnotation(AnnotationSpec.builder(JsonDeserialize.class)
+                        .addMember("builder", "$T.class", builderImplementation)
+                        .build());
                 typeBuilder
                         .addMethod(createBuilder(builderClass))
                         .addType(BeanBuilderGenerator.generate(
@@ -177,60 +171,15 @@ public final class BeanGenerator {
                                 options,
                                 Optional.empty()));
             } else {
-                List<TypeSpec> interfaces = generateStageInterfaces(
+                BeanBuilderGenerator.generateStagedBuilder(
+                        typeBuilder,
+                        typeMapper,
+                        safetyEvaluator,
                         objectClass,
                         builderClass,
-                        typeMapper,
-                        fieldsNeedingBuilderStage,
-                        fields.stream()
-                                .filter(BeanGenerator::fieldShouldBeInFinalStage)
-                                .collect(ImmutableList.toImmutableList()),
-                        safetyEvaluator);
-
-                List<ClassName> interfacesAsClasses = interfaces.stream()
-                        .map(stageInterface ->
-                                ClassName.get(objectClass.packageName(), objectClass.simpleName(), stageInterface.name))
-                        .collect(Collectors.toList());
-
-                TypeSpec builderInterface = TypeSpec.interfaceBuilder("Builder")
-                        .addModifiers(Modifier.PUBLIC)
-                        .addMethods(interfaces.stream()
-                                .map(stageInterface -> stageInterface.methodSpecs)
-                                .flatMap(List::stream)
-                                .map(method -> method.toBuilder()
-                                        .addAnnotation(Override.class)
-                                        .returns(
-                                                method.name.equals("build")
-                                                        ? method.returnType
-                                                        : ClassName.get(
-                                                                objectClass.packageName(),
-                                                                objectClass.simpleName(),
-                                                                "Builder"))
-                                        .build())
-                                .collect(Collectors.toList()))
-                        .addSuperinterfaces(interfacesAsClasses.stream()
-                                .map(Primitives::box)
-                                .collect(Collectors.toList()))
-                        .build();
-
-                typeBuilder
-                        .addTypes(interfaces)
-                        .addType(builderInterface)
-                        .addMethod(MethodSpec.methodBuilder("builder")
-                                .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                                .returns(interfacesAsClasses.get(0))
-                                .addStatement("return new DefaultBuilder()")
-                                .build())
-                        .addType(BeanBuilderGenerator.generate(
-                                typeMapper,
-                                safetyEvaluator,
-                                objectClass,
-                                builderClass,
-                                typeDef,
-                                typesMap,
-                                options,
-                                Optional.of(ClassName.get(
-                                        objectClass.packageName(), objectClass.simpleName(), builderInterface.name))));
+                        typeDef,
+                        typesMap,
+                        options);
             }
         }
         typeBuilder.addAnnotation(ConjureAnnotations.getConjureGeneratedAnnotation(BeanGenerator.class));
@@ -241,153 +190,6 @@ public final class BeanGenerator {
                 .skipJavaLangImports(true)
                 .indent("    ")
                 .build();
-    }
-
-    private static boolean fieldShouldBeInFinalStage(EnrichedField field) {
-        Type type = field.conjureDef().getType();
-        return type.accept(TypeVisitor.IS_LIST)
-                || type.accept(TypeVisitor.IS_SET)
-                || type.accept(TypeVisitor.IS_MAP)
-                || type.accept(TypeVisitor.IS_OPTIONAL);
-    }
-
-    /**
-     * Sorts input fields in the order they should be applied to the builder: Original order with required
-     * fields prior to optional/collection values.
-     */
-    private static Stream<EnrichedField> sortedEnrichedFields(ImmutableList<EnrichedField> enrichedFields) {
-        return enrichedFields.stream()
-                .sorted(Comparator.comparing(BeanGenerator::fieldShouldBeInFinalStage)
-                        .thenComparing(enrichedFields::indexOf));
-    }
-
-    private static ClassName stageBuilderInterfaceName(ClassName enclosingClass, String stageName) {
-        return enclosingClass.nestedClass(StringUtils.capitalize(stageName) + "StageBuilder");
-    }
-
-    private static List<TypeSpec> generateStageInterfaces(
-            ClassName objectClass,
-            ClassName builderClass,
-            TypeMapper typeMapper,
-            ImmutableList<EnrichedField> fieldsNeedingBuilderStage,
-            ImmutableList<EnrichedField> otherFields,
-            SafetyEvaluator safetyEvaluator) {
-        List<TypeSpec.Builder> interfaces = new ArrayList<>();
-
-        PeekingIterator<EnrichedField> fieldPeekingIterator = Iterators.peekingIterator(
-                sortedEnrichedFields(fieldsNeedingBuilderStage).iterator());
-
-        while (fieldPeekingIterator.hasNext()) {
-            EnrichedField field = fieldPeekingIterator.next();
-            String nextBuilderStageName = fieldPeekingIterator.hasNext()
-                    ? JavaNameSanitizer.sanitize(fieldPeekingIterator.peek().fieldName())
-                    : "completed_";
-
-            ClassName nextStageClassName = stageBuilderInterfaceName(objectClass, nextBuilderStageName);
-
-            interfaces.add(TypeSpec.interfaceBuilder(
-                            stageBuilderInterfaceName(objectClass, JavaNameSanitizer.sanitize(field.fieldName())))
-                    .addModifiers(Modifier.PUBLIC)
-                    .addMethod(MethodSpec.methodBuilder(JavaNameSanitizer.sanitize(field.fieldName()))
-                            .addParameter(ParameterSpec.builder(
-                                            field.poetSpec().type, JavaNameSanitizer.sanitize(field.fieldName()))
-                                    .addAnnotation(Nonnull.class)
-                                    .build())
-                            .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
-                            .returns(Primitives.box(nextStageClassName))
-                            .build()));
-        }
-
-        ClassName completedStageClass = stageBuilderInterfaceName(objectClass, "completed_");
-
-        TypeSpec.Builder completedStage = TypeSpec.interfaceBuilder(completedStageClass)
-                .addModifiers(Modifier.PUBLIC)
-                .addMethod(MethodSpec.methodBuilder("build")
-                        .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
-                        .returns(Primitives.box(objectClass))
-                        .build());
-
-        completedStage.addMethods(otherFields.stream()
-                .map(field ->
-                        generateMethodsForFinalStageField(field, typeMapper, completedStageClass, safetyEvaluator))
-                .flatMap(List::stream)
-                .collect(Collectors.toList()));
-
-        interfaces.add(completedStage);
-
-        interfaces
-                .get(0)
-                .addMethod(MethodSpec.methodBuilder("from")
-                        .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
-                        .returns(builderClass)
-                        .addParameter(objectClass, "other")
-                        .build());
-
-        return interfaces.stream().map(TypeSpec.Builder::build).collect(Collectors.toList());
-    }
-
-    private static List<MethodSpec> generateMethodsForFinalStageField(
-            EnrichedField enriched, TypeMapper typeMapper, ClassName returnClass, SafetyEvaluator safetyEvaluator) {
-        List<MethodSpec> methodSpecs = new ArrayList<>();
-        Type type = enriched.conjureDef().getType();
-        FieldDefinition definition = enriched.conjureDef();
-        Optional<LogSafety> safety = safetyEvaluator.getUsageTimeSafety(definition);
-
-        methodSpecs.add(MethodSpec.methodBuilder(JavaNameSanitizer.sanitize(enriched.fieldName()))
-                .addParameter(ParameterSpec.builder(
-                                BeanBuilderAuxiliarySettersUtils.widenParameterIfPossible(
-                                        enriched.poetSpec().type, type, typeMapper, safety),
-                                JavaNameSanitizer.sanitize(enriched.fieldName()))
-                        .addAnnotation(Nonnull.class)
-                        .build())
-                .addJavadoc(Javadoc.render(definition.getDocs(), definition.getDeprecated())
-                        .map(rendered -> CodeBlock.of("$L", rendered))
-                        .orElseGet(() -> CodeBlock.builder().build()))
-                .addAnnotations(ConjureAnnotations.deprecation(definition.getDeprecated()))
-                .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
-                .returns(Primitives.box(returnClass))
-                .build());
-
-        if (type.accept(TypeVisitor.IS_LIST)) {
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createCollectionSetterBuilder(
-                            "addAll", enriched, typeMapper, returnClass, safetyEvaluator)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createItemSetterBuilder(
-                            enriched, type.accept(TypeVisitor.LIST).getItemType(), typeMapper, returnClass, safety)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-        }
-
-        if (type.accept(TypeVisitor.IS_SET)) {
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createCollectionSetterBuilder(
-                            "addAll", enriched, typeMapper, returnClass, safetyEvaluator)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createItemSetterBuilder(
-                            enriched, type.accept(TypeVisitor.SET).getItemType(), typeMapper, returnClass, safety)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-        }
-
-        if (type.accept(TypeVisitor.IS_MAP)) {
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createCollectionSetterBuilder(
-                            "putAll", enriched, typeMapper, returnClass, safetyEvaluator)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createMapSetterBuilder(enriched, typeMapper, returnClass)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-        }
-
-        if (type.accept(TypeVisitor.IS_OPTIONAL)) {
-            methodSpecs.add(BeanBuilderAuxiliarySettersUtils.createOptionalSetterBuilder(
-                            enriched, typeMapper, returnClass, safetyEvaluator)
-                    .addModifiers(Modifier.ABSTRACT)
-                    .build());
-        }
-
-        return methodSpecs;
     }
 
     private static boolean useCachedHashCode(Collection<EnrichedField> fields) {
@@ -557,13 +359,15 @@ public final class BeanGenerator {
                     .addAnnotations(ConjureAnnotations.safety(safetyEvaluator.getUsageTimeSafety(field.conjureDef())))
                     .build()));
             // Follow order on adding methods on builder to comply with staged builders option if set
-            sortedEnrichedFields(fields).map(EnrichedField::poetSpec).forEach(spec -> {
-                if (isOptional(spec)) {
-                    builder.addCode("\n    .$L(Optional.of($L))", spec.name, spec.name);
-                } else {
-                    builder.addCode("\n    .$L($L)", spec.name, spec.name);
-                }
-            });
+            BeanBuilderGenerator.sortedEnrichedFields(fields)
+                    .map(EnrichedField::poetSpec)
+                    .forEach(spec -> {
+                        if (isOptional(spec)) {
+                            builder.addCode("\n    .$L(Optional.of($L))", spec.name, spec.name);
+                        } else {
+                            builder.addCode("\n    .$L($L)", spec.name, spec.name);
+                        }
+                    });
             builder.addCode("\n    .build();\n");
         }
